Compute face properties.
Split the face properties out of the Cell class into their own.
Store these Faces as a vector in a map indexed by the direction
of the plane normal vector. Each Face stores this same normal vector,
its centroid and area. It also stores a boolean value indicating
whether it's neighbour on the grid is a fluid cell (as opposed to a
purely solid cell).

Still to do:
- Compute the boolean value 'fluid'.
- Compute the cell volume.