Implement cut cell output.
Output cut cells in CGNS format with the "TETRA_4" element type. This involves
tetrahedralising the arbitrary Nef_polyhedron created in the cutting stage. To
do this, we use the CGAL::Delaunay_triangulation_3 class. This class does not
handle concave polyhedra, so we use the CGAL::convex_decomposition_3 class to
ensure that each polyhedra is tetrahedralisable.

Also, convert the fluid output routine to output HEXA_8 elements instead of
arbitrary polyhedra. This is a lot simpler when everything is a cube.