Drastically improve memory usage.
Move the creation of the unit cube at each point into the cutting routine. This
dramatically reduces the amount of memory needed from scaling with the size of
the grid to scaling with the size of the solid object.

One important side-effect is that the Nef_polyhedron at each point in the N_
array is only defined if the cell type is Cut. This shouldn't be a problem, as
it is trivial to recreate the Fluid or Solid cells (they are the UnitCube
suitably translated and the empty Nef_polyhedron, respectively) if required.

If it proves to be necessary, the memory usage could be reduced even further,
such that only one Nef_polyhedron needs to be stored at any moment. To achieve
this, the Grid::cut and output functions should be combined. The output
functions could be converted to classes with callback functions to start and
stop output, and a callback function for each point in the grid. In this case,
it may prove useful to have Grid::cut accept a vector<Nef_polyhedron>, as well.
However, this may be unnecessary, as the caller can already combine many solids
easily using the union operation on Nef_polyhedra.