Optimise creating the list of grid coordinates.
The naive map implementation used to create the CGNS grid coordinate lists
appears to be a bottleneck. Since Fluid cells mostly share cells with their
neighbours, which have been recently added, it makes sense to scan the list of
points backwards.

This slightly complicates the computation of the index of points within these
lists, so the slower algorithm has been maintained and will be used to test
the result of the optimised algorithm if NDEBUG is not defined (e.g. when using
CMake's Debug build type).