Optimise by pre-allocating CGNS arrays.
This does prevent allocations in the critical sections of the code, but it's
unclear if this actually saves time. It may provide a nicer failure mode if
memory is tight. That is, on systems which check allocations before use (not
Linux in the default configuration) the program can fail immediately instead
of dying after potentially doing a lot of work.

Convert a list to a vector.

Also add a lot of debugging output.