Add boundary faces.
Write a CGNS Elements section for QUAD_4 faces on the bounding box. Write a
CGNS BC_t (Boundary Conditions) section for each of the six bounding faces..

This code only writes faces and boundary conditions for the faces on the
bounding box. The code currently only works for grids which have only Fluid
cells on the bounding box, and will ignore the Cut cells in production mode and
assert() in debug mode if a Cut or Solid cell is on the boundary.