Convert find_or_insert_index_of to a class. Add an assert.
It makes much more sense to store the three double vectors and the Point_3
vector in a class, so that we don't have to continually pass them around.

Also assert if we end up having to create a new Point_3 _after_ the
GridCoordinates have been written to the CGNS file. If this happens, the new
point will never be written and the element reference will be invalid.