Replace direct widget access with a signal.
It is not safe to alter the state of the GUI QMainWindow from the QThread which
is doing the computation. Often the GUI would freeze. To fix this, create a
slot on the main window and connect a signal in the thread to signal
completion. QT's automatic connection will make the call in the thread of the
receiver, avoiding the need for locking.

There may still be a need to lock accesses to the data of the widgets within
the QMainWindow, but it doesn't seem to be a problem yet. If so, one solution
is to copy all the data to the GenerateThread object before starting it.